requirements.txt
fastapi==0.115.12
pandas==2.2.3
psycopg2_binary==2.9.10
pydantic==2.11.4
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.0
transformers==4.51.3
